feat: change the name temp provider to Provider
